# Conflicts:
#	Dockerfile
#	Jenkinsfile
#	nemo/collections/asr/parts/submodules/rnnt_beam_decoding.py
#	nemo/collections/nlp/models/duplex_text_normalization/duplex_decoder.py
#	nemo/collections/tts/helpers/helpers.py
#	nemo/collections/tts/modules/waveglow.py
#	nemo/package_info.py
#	nemo_text_processing/text_normalization/normalize_with_audio.py
#	tools/nmt_grpc_service/server.py
#	tutorials/asr/ASR_CTC_Language_Finetuning.ipynb
#	tutorials/asr/ASR_with_Subword_Tokenization.ipynb
#	tutorials/asr/ASR_with_Transducers.ipynb
#	tutorials/asr/Intro_to_Transducers.ipynb
#	tutorials/nlp/01_Pretrained_Language_Models_for_Downstream_Tasks.ipynb
#	tutorials/nlp/Entity_Linking_Medical.ipynb
#	tutorials/nlp/Non_English_Downstream_Tasks_(NER).ipynb
#	tutorials/nlp/Punctuation_and_Capitalization.ipynb
#	tutorials/nlp/Relation_Extraction-BioMegatron.ipynb
#	tutorials/nlp/Token_Classification-BioMegatron.ipynb
#	tutorials/speaker_recognition/ASR_with_SpeakerDiarization.ipynb
#	tutorials/tools/CTC_Segmentation_Tutorial.ipynb
#	tutorials/tts/Tacotron2_Training.ipynb

+++ examples/asr/transcribe_speech.py
@@ -30,23 +30,39 @@ from nemo.utils import logging, model_utils
 
 
 """
-# Transcribe audio
+Transcribe audio file on a single CPU/GPU. Useful for transcription of moderate amounts of audio data.
+
 # Arguments
-# model_path: path to .nemo ASR checkpoint
-# pretrained_name: name of pretrained ASR model (from NGC registry)
-# audio_dir: path to directory with audio files
-# dataset_manifest: path to dataset JSON manifest file (in NeMo format)
-#
-# ASR model can be specified by either "model_path" or "pretrained_name".
-# Data for transcription can be defined with either "audio_dir" or "dataset_manifest".
-# Results are returned in a JSON manifest file.
+  model_path: path to .nemo ASR checkpoint
+  pretrained_name: name of pretrained ASR model (from NGC registry)
+  audio_dir: path to directory with audio files
+  dataset_manifest: path to dataset JSON manifest file (in NeMo format)
+  
+  output_filename: Output filename where the transcriptions will be written
+  batch_size: batch size during inference
+  
+  cuda: Optional int to enable or disable execution of model on certain CUDA device.
+  amp: Bool to decide if Automatic Mixed Precision should be used during inference
+  audio_type: Str filetype of the audio. Supported = wav, flac, mp3
+  
+  overwrite_transcripts: Bool which when set allowes repeated transcriptions to overwrite previous results.
+  
+  rnnt_decoding: Decoding sub-config for RNNT. Refer to documentation for specific values.
+
+# Usage
+ASR model can be specified by either "model_path" or "pretrained_name".
+Data for transcription can be defined with either "audio_dir" or "dataset_manifest".
+Results are returned in a JSON manifest file.
 
 python transcribe_speech.py \
     model_path=null \
     pretrained_name=null \
     audio_dir="" \
     dataset_manifest="" \
-    output_filename=""
+    output_filename="" \
+    batch_size=32 \
+    cuda=0 \
+    amp=True
 """
 
 
@@ -61,16 +77,24 @@ class TranscriptionConfig:
     # General configs
     output_filename: Optional[str] = None
     batch_size: int = 32
-    cuda: Optional[bool] = None  # will switch to cuda if available, defaults to CPU otherwise
+    num_workers: int = min(batch_size, os.cpu_count() - 1)
+
+    # Set `cuda` to int to define CUDA device. If 'None', will look for CUDA
+    # device anyway, and do inference on CPU only if CUDA device is not found.
+    # If `cuda` is a negative number, inference will be on CPU only.
+    cuda: Optional[int] = None
     amp: bool = False
     audio_type: str = "wav"
 
-    # decoding strategy for RNNT models
+    # Recompute model transcription, even if the output folder exists with scores.
+    overwrite_transcripts: bool = True
+
+    # Decoding strategy for RNNT models
     rnnt_decoding: RNNTDecodingConfig = RNNTDecodingConfig()
 
 
 @hydra_runner(config_name="TranscriptionConfig", schema=TranscriptionConfig)
-def main(cfg: TranscriptionConfig):
+def main(cfg: TranscriptionConfig) -> TranscriptionConfig:
     logging.info(f'Hydra config: {OmegaConf.to_yaml(cfg)}')
 
     if cfg.model_path is None and cfg.pretrained_name is None:
@@ -80,14 +104,12 @@ def main(cfg: TranscriptionConfig):
 
     # setup GPU
     if cfg.cuda is None:
-        cfg.cuda = torch.cuda.is_available()
+        if torch.cuda.is_available():
+            cfg.cuda = 0  # use 0th CUDA device
+        else:
+            cfg.cuda = -1  # use CPU
 
-    if type(cfg.cuda) == int:
-        device_id = int(cfg.cuda)
-    else:
-        device_id = 0
-
-    device = torch.device(f'cuda:{device_id}' if cfg.cuda else 'cpu')
+    device = torch.device(f'cuda:{cfg.cuda}' if cfg.cuda >= 0 else 'cpu')
 
     # setup model
     if cfg.model_path is not None:
@@ -103,7 +125,7 @@ def main(cfg: TranscriptionConfig):
         asr_model = ASRModel.from_pretrained(model_name=cfg.pretrained_name, map_location=device)  # type: ASRModel
         model_name = cfg.pretrained_name
 
-    trainer = pl.Trainer(gpus=int(cfg.cuda))
+    trainer = pl.Trainer(gpus=[cfg.cuda] if cfg.cuda >= 0 else 0)
     asr_model.set_trainer(trainer)
     asr_model = asr_model.eval()
 
@@ -133,12 +155,7 @@ def main(cfg: TranscriptionConfig):
         def autocast():
             yield
 
-    # transcribe audio
-    with autocast():
-        with torch.no_grad():
-            transcriptions = asr_model.transcribe(filepaths, batch_size=cfg.batch_size)
-    logging.info(f"Finished transcribing {len(filepaths)} files !")
-
+    # Compute output filename
     if cfg.output_filename is None:
         # create default output filename
         if cfg.audio_dir is not None:
@@ -146,12 +163,28 @@ def main(cfg: TranscriptionConfig):
         else:
             cfg.output_filename = cfg.dataset_manifest.replace('.json', f'_{model_name}.json')
 
+    # if transcripts should not be overwritten, and already exists, skip re-transcription step and return
+    if not cfg.overwrite_transcripts and os.path.exists(cfg.output_filename):
+        logging.info(
+            f"Previous transcripts found at {cfg.output_filename}, and flag `overwrite_transcripts`"
+            f"is {cfg.overwrite_transcripts}. Returning without re-transcribing text."
+        )
+
+        return cfg
+
+    # transcribe audio
+    with autocast():
+        with torch.no_grad():
+            transcriptions = asr_model.transcribe(filepaths, batch_size=cfg.batch_size)
+    logging.info(f"Finished transcribing {len(filepaths)} files !")
+
     logging.info(f"Writing transcriptions into file: {cfg.output_filename}")
 
     # if transcriptions form a tuple (from RNNT), extract just "best" hypothesis
     if type(transcriptions) == tuple and len(transcriptions) == 2:
         transcriptions = transcriptions[0]
 
+    # write audio transcriptions
     with open(cfg.output_filename, 'w', encoding='utf-8') as f:
         if cfg.audio_dir is not None:
             for idx, text in enumerate(transcriptions):
@@ -165,6 +198,7 @@ def main(cfg: TranscriptionConfig):
                     f.write(json.dumps(item) + "\n")
 
     logging.info("Finished writing predictions !")
+    return cfg
 
 
 if __name__ == '__main__':
